Item Cards are ready to be well displayed
