+++ elhura-ecom/src/components/ArticlesManegement/ArticleItem.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
+//import { makeStyles } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
 import CardActions from '@material-ui/core/CardActions';
 import CardContent from '@material-ui/core/CardContent';
@@ -8,7 +8,7 @@ import Typography from '@material-ui/core/Typography';
 import { Avatar, CardHeader, CardMedia, IconButton } from '@material-ui/core';
 import ShareIcon from '@material-ui/icons/Share';
 
-const useStyles = makeStyles({
+/* const useStyles = makeStyles({
 
     bullet: {
         display: 'inline-block',
@@ -21,11 +21,10 @@ const useStyles = makeStyles({
     pos: {
         marginBottom: 12,
     },
-});
+}); */
 
 const ArticleItem = (props) => {
-    const classes = useStyles();
-    const bull = <span className={classes.bullet}>•</span>;
+    // const classes = useStyles();
     const { avatarUrl, title, subtitle, description, imageUrl } = props;
