+++ elhura-ecom/src/components/ArticlesManegement/Content.jsx
@@ -1,0 +1,24 @@
+  
+import React from "react";
+import ArticleItem from "./ArticleItem"
+import { Grid } from "@material-ui/core";
+import articles from '../../mockData/constants'
+
+const Content = () => {
+    const getAricles = (aricle) => {
+        return (
+            <Grid item xs={12} sm={4}>
+                <ArticleItem {...aricle} />
+            </Grid>
+        );
+    }
+
+
+  return (
+    <Grid container spacing={2}>
+      {articles.map(artcile => getAricles(artcile))}
+    </Grid>
+  );
+};
+
+export default Content;
